requirements.txt
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
drf-writable-nested==0.7.0
gunicorn==23.0.0
pillow==10.4.0
psycopg2-binary==2.9.9
PyPDF2==3.0.1 # test only
python-dotenv==1.0.1
weasyprint==62.3
